Fix closing brackets when a space is present.
The standard behavior of AutoClose is to skip the closing delimiter
when it is present after the current cursor position. I usually have
whitespace in between the cursor position and the closing delimiter.
Now the cursor is placed after the closing delimiter when only
whitespace separates the cursor and the delimiter.