Stop printing errors when inside a .git folder.
Normally when we are not in a git repository `git branch` will fail.
In which case there is no point in running other git related
commands, and they can't "fail". However if we are inside a .git
folder the `git branch` command succeeds but `git status --porcelain`
will cause error output. This is fixed by writing stderr of `git
status` to /dev/null as well.