Replace tmux leader to <C-z> instead of <C-t>
The 't' is used for navigations or directions. By making it part of the
tmux leader it is no longer available in combination with the control
modifier.

In vim '<C-h>' should be mapped to scroll a page up, and '<C-t>' would
be scroll a page down. But that is currently not possible when using vim
within a tmux session.